This commit was generated by cvs2svn to compensate for changes in r158, which included commits to RCS files with non-trunk default branches.
git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@159 06663e23-700e-0410-b217-a244a6096597

+++ pcl/pgstate.h
@@ -206,7 +206,7 @@ typedef struct pcl_hpgl_state_s {
 	  struct fp_ {
 	    hpgl_hatch_params_t hatch;
 	    hpgl_hatch_params_t crosshatch;
-	    float shading;
+	    int shading;	/* 0..100 */
 	    int pattern_index;
 	    int pattern_type;
 	    uint pattern_id;
